requirements/requirements.txt
torch
numpy
matplotlib
opencv_python
tensorboardX
PyYAML
pycocotools
